[20001] : Conversion de certains documents pdf dans le dossier d'impression (develop)

See merge request maarch/MaarchCourrier!619

+++ src/app/resource/controllers/FolderPrintController.php
@@ -561,6 +561,7 @@ class FolderPrintController
                 $documentPaths = array_merge($documentPaths, $linkedAttachmentPath);
             }
 
+
             if (!empty($documentPaths)) {
                 if (empty($resource['altIdentifier'] . $resource['subject'])) {
                     $document = ResModel::getById([
@@ -589,8 +590,8 @@ class FolderPrintController
                 if (file_exists($filePathOnTmp)) {
                     unlink($filePathOnTmp);
                 }
-                $command = "pdfunite '" . implode("' '", $documentPaths) . "' '" . $filePathOnTmp . "'";
 
+                $command = "pdfunite '" . implode("' '", $documentPaths) . "' '" . $filePathOnTmp . "'";
                 exec($command . ' 2>&1', $output, $return);
 
                 if (!file_exists($filePathOnTmp)) {
@@ -599,7 +600,10 @@ class FolderPrintController
 
                 // delete all tmp email_*.pdf, attachment_*.pdf, summarySheet_*.pdf, convertedAr_*.pdf and listNotes_*.pdf after merged is complete
                 foreach ($documentPaths as $documentPath) {
-                    unlink($documentPath);
+                    if (strpos($documentPath, "email_") !== false           || strpos($documentPath, "attachment_") !== false   || strpos($documentPath, "summarySheet_") !== false 
+                        || strpos($documentPath, "convertedAr_") !== false  || strpos($documentPath, "listNotes_") !== false) {
+                        unlink($documentPath);
+                    }
                 }
                 
                 $folderPrintPaths[] = $filePathOnTmp;
@@ -616,42 +620,40 @@ class FolderPrintController
 
             $response = $response->withAddedHeader('Content-Disposition', "inline; filename=maarch.pdf");
             return $response->withHeader('Content-Type', $mimeType);
-        } else {
-            $filePathOnTmp = str_replace('//', '/', $tmpDir) . 'folderPrint.zip';
-            if (file_exists($filePathOnTmp)) {
-                unlink($filePathOnTmp);
-            }
-
-            $zip = new \ZipArchive;
-            if ($zip->open($filePathOnTmp, \ZipArchive::CREATE) !== TRUE) {
-                return $response->withStatus(500)->withJson(['errors' => 'Merged ZIP file not created']);
-            }
-            foreach ($folderPrintPaths as $folderPrintPath) {
-                $zip->addFile($folderPrintPath, basename($folderPrintPath));
-            }
-            $zip->close();
-
-            if (!file_exists($filePathOnTmp)) {
-                return $response->withStatus(500)->withJson(['errors' => 'Merged ZIP file not created']);
-            }
-
-            $finfo = new \finfo(FILEINFO_MIME_TYPE);
-            $fileContent = file_get_contents($filePathOnTmp);
-            $mimeType = $finfo->buffer($fileContent);
-
-            $response->write($fileContent);
-
-            // delete tmp files, partly to avoid filling an existing ZIP and sending more than was requested
-            unlink($filePathOnTmp);
-            foreach ($folderPrintPaths as $folderPrintPath) {
-                unlink($folderPrintPath);
-            }
-
-            $response = $response->withAddedHeader('Content-Disposition', 'inline; filename=maarch.zip');
-            return $response->withHeader('Content-Type', $mimeType);
         }
 
-        return $response->withStatus(400)->withJson(['errors' => 'No document to merge']);
+        $filePathOnTmp = str_replace('//', '/', $tmpDir) . 'folderPrint.zip';
+        if (file_exists($filePathOnTmp)) {
+            unlink($filePathOnTmp);
+        }
+
+        $zip = new \ZipArchive;
+        if ($zip->open($filePathOnTmp, \ZipArchive::CREATE) !== TRUE) {
+            return $response->withStatus(500)->withJson(['errors' => 'Merged ZIP file not created']);
+        }
+        foreach ($folderPrintPaths as $folderPrintPath) {
+            $zip->addFile($folderPrintPath, basename($folderPrintPath));
+        }
+        $zip->close();
+
+        if (!file_exists($filePathOnTmp)) {
+            return $response->withStatus(500)->withJson(['errors' => 'Merged ZIP file not created']);
+        }
+
+        $finfo = new \finfo(FILEINFO_MIME_TYPE);
+        $fileContent = file_get_contents($filePathOnTmp);
+        $mimeType = $finfo->buffer($fileContent);
+
+        $response->write($fileContent);
+
+        // delete tmp files, partly to avoid filling an existing ZIP and sending more than was requested
+        unlink($filePathOnTmp);
+        foreach ($folderPrintPaths as $folderPrintPath) {
+            unlink($folderPrintPath);
+        }
+
+        $response = $response->withAddedHeader('Content-Disposition', 'inline; filename=maarch.zip');
+        return $response->withHeader('Content-Type', $mimeType);
     }
 
     private static function getDocumentFilePath(array $args)
